Rearrange tests to mirror application organization
