--- spec/support/jasmine.json
@@ -1,7 +1,0 @@
-{
-  "spec_dir": "spec",
-  "spec_files": ["**/*[sS]pec.ts"],
-  "helpers": ["helpers/**/*.js"],
-  "stopSpecOnExpectationFailure": false,
-  "random": true
-}
