+++ src/test/mocks/vscode.mock.ts
@@ -1,12 +1,11 @@
 export class VscodeMock {
-
- workspace: {
+ workspace = {
     getConfiguration() {
       return {
         get() {
-          return 'Daily'
+          return 'Daily';
         },
       }
     }
-  },
+  };
 }
