+++ src/test/suite/common/utils.test.ts
@@ -1,0 +1,18 @@
+import * as assert from 'assert'
+
+import { Util } from '../../../common/util'
+
+suite('Util', async () => {
+  suite('userCancelled()', () => {
+    test('returns true when passed "Cancel"', () => {
+      const result = Util.userCancelled('Cancel')
+
+      assert.equal(result, true)
+    })
+    test('returns false when not "Cancel" is passed in', () => {
+      const isFalse = Util.userCancelled('aldfjalsfj')
+
+      assert.equal(isFalse, false)
+    })
+  })
+})
