+++ src/test/suite/extension.test.ts
@@ -6,11 +6,6 @@
 // The module 'assert' provides assertion methods from node
 import * as assert from 'assert'
 
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
-// import * as vscode from 'vscode';
-// import * as myExtension from '../extension';
-
 // Defines a Mocha test suite to group tests of similar kind together
 suite('Extension Tests', function() {
   // Defines a Mocha unit test
